requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
greenlet==3.0.1
openai==1.3.0
requests==2.31.0
torch==2.0.1
transformers==4.30.2
accelerate==0.20.3